add: parse connections and add them to nodes

+++ connection.zig
@@ -2,7 +2,7 @@ const Node = @import("node.zig").Node;
 
 pub const Connection = struct {
     target: *Node,
-    drive_time: u16,
-    distance: u16,
+    drive_time: u32,
+    distance: u32,
     speed_limit: u8,
 };
